Finished scene transfer, tweaked HotelPresent to account for new BG
The player can now move between scenes at specified points. And upon changing scenes, the previous position is saved to that it can be used instead. The camera also doesn't catch up at the start of the scene, it instantly fixes itself on the player and then returns to normal smoothing.